refactor: manager-cart feature typescript 변경
