--- src/features/manage-cart/lib.js
@@ -1,2 +1,0 @@
-export const generateCartItemText = (product, quantityInCart) =>
-  `${product.name} - ${product.price}원 x ${quantityInCart}`;
